+++ pywrap/templates.py
@@ -38,14 +38,14 @@ template_function_decl = """cdef extern from "%(filename)s" namespace "%(namespa
 method_decl = "%(result_type)s %(name)s(%(args)s)"
 template_method_decl = "%(result_type)s %(name)s[%(types)s](%(args)s)"
 constructor_decl = "%(class_name)s(%(args)s)"
-arg_decl = "%(tipe)s %(name)s"
+arg_decl = "%(type_prefix)s%(tipe)s %(name)s"
 field_decl = "%(tipe)s %(name)s"
 
 # member definitions
 ctor_default_def = """    def __init__(cpp.%(name)s self):
-        self.thisptr = new cpp.%(name)s()"""
-fun_call = "cpp.%(name)s(%(call_args)s)"
-ctor_call = "self.thisptr = new cpp.%(class_name)s(%(call_args)s)"
+        self.thisptr = new _%(modulename)s.%(name)s()"""
+fun_call = "_%(modulename)s.%(name)s(%(call_args)s)"
+ctor_call = "self.thisptr = new _%(modulename)s.%(class_name)s(%(call_args)s)"
 method_call = "self.thisptr.%(name)s(%(call_args)s)"
 setter_call = "self.thisptr.%(name)s = %(call_args)s"
 getter_call = "self.thisptr.%(name)s"